+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.19.2 (Unreleased)
 ENHANCEMENTS:
+* Suppress diffs for equivalent values of some time attributes. This should reduce output of `terraform plan` that didn't disappear after running `terraform apply`. (GH-442)
 * Add better validation and documentation for some time attributes (GH-443)
 
 ## 2.19.1 (October 6, 2022)